requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers>=4.43.0
tokenizers>=0.15.2
datasets>=2.20.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
loguru>=0.7.2
tensorboard>=2.16.0
pyyaml>=6.0.1
einops>=0.7.0
scipy>=1.12.0
sentencepiece>=0.1.99